requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-redis==4.11.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
pytz==2019.3
redis==3.3.11
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==2.1.0
